fix linspace

+++ thorns/waves.py
@@ -75,7 +75,7 @@ def fft_filter(signal, fs, band):
 
     lo, hi = band
 
-    freqs = np.linspace(0, fs/2, len(signal)/2+1)
+    freqs = np.linspace(0, fs/2, int(len(signal)/2 + 1))
 
     signal_fft = np.fft.rfft(signal)
     signal_fft[(freqs < lo) | (freqs > hi)] = 0
@@ -168,7 +168,7 @@ def ramped_tone(
         s = set_dbspl(s, dbspl)
 
     if ramp != 0:
-        ramp_signal = np.linspace(0, 1, np.ceil(ramp * fs))
+        ramp_signal = np.linspace(0, 1, int(np.ceil(ramp * fs)))
         s[0:len(ramp_signal)] = s[0:len(ramp_signal)] * ramp_signal
         s[-len(ramp_signal):] = s[-len(ramp_signal):] * ramp_signal[::-1]
 
@@ -211,7 +211,7 @@ def white_noise(
     s = set_dbspl(s, dbspl)
 
     # Ramping
-    ramp_signal = np.linspace(0, 1, np.round(ramp*fs))
+    ramp_signal = np.linspace(0, 1, int(np.round(ramp * fs)))
     s[0:len(ramp_signal)] = s[0:len(ramp_signal)] * ramp_signal
     s[-len(ramp_signal):] = s[-len(ramp_signal):] * ramp_signal[::-1]
 
@@ -408,7 +408,7 @@ def amplitude_modulated_tone(
         s = set_dbspl(s, dbspl)
 
     if ramp != 0:
-        ramp_signal = np.linspace(0, 1, np.ceil(ramp * fs))
+        ramp_signal = np.linspace(0, 1, int(np.ceil(ramp * fs)))
         s[0:len(ramp_signal)] = s[0:len(ramp_signal)] * ramp_signal
         s[-len(ramp_signal):] = s[-len(ramp_signal):] * ramp_signal[::-1]
